+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## Unreleased
+### Added
+* It is now possible to validate multiple messages at once by adding a parameter to the `send*` Methods
+  ([Documentation](https://firebase-php.readthedocs.io/en/latest/cloud-messaging.html#validating-messages))
+* It is now possible to check a list of registration tokens whether they are valid and known, unknown, or invalid
+  ([Documentation](https://firebase-php.readthedocs.io/en/latest/cloud-messaging.html#validating-registration-tokens))
+* Added methods:
+  * `Kreait\Firebase\Messaging::validateRegistrationTokens($registrationTokenOrTokens)`
+
 ## [5.13.0] - 2020-12-10
 
 This release ensures compatibility with PHP 8.0